# docs/swagger.yaml
basePath: /api
definitions:
  chinese-game-backend_internal_domain.DialogStepItem:
    properties:
      bg:
        type: string
      emotion:
        type: string
      speaker:
        type: string
      text:
        type: string
    type: object
  chinese-game-backend_internal_domain.Level:
    properties:
      background_src:
        type: string
      color:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      order_index:
        type: integer
      planet_img_src:
        type: string
      steps:
        items:
          $ref: '#/definitions/chinese-game-backend_internal_domain.LevelStep'
        type: array
      title:
        type: string
    type: object
  chinese-game-backend_internal_domain.LevelStep:
    properties:
      content:
        items:
          type: integer
        type: array
      description:
        type: string
      dialog:
        items:
          $ref: '#/definitions/chinese-game-backend_internal_domain.DialogStepItem'
        type: array
      id:
        type: integer
      is_completed:
        type: boolean
      level_id:
        type: integer
      order_index:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/chinese-game-backend_internal_domain.LevelStepType'
    type: object
  chinese-game-backend_internal_domain.LevelStepType:
    enum:
    - theory
    - quiz
    - final
    - dialog
    - categorization
    - tone_listening
    - planet_click
    - planet_matching
    type: string
    x-enum-varnames:
    - LevelStepThoery
    - LevelStepQuiz
    - LevelStepFinal
    - LevelStepDialog
    - LevelStepCategoriz
    - LevelStepToneListen
    - LevelStepPlanetClick
    - LevelStepPlanetMatch
  internal_transport_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_transport_http.completeStepInput:
    properties:
      step_id:
        type: integer
    type: object
  internal_transport_http.joinStudentInput:
    properties:
      invite_code:
        type: string
    type: object
  internal_transport_http.loginResponse:
    properties:
      token:
        type: string
    type: object
  internal_transport_http.signUpInput:
    properties:
      invite_code:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  internal_transport_http.studentProgressResponse:
    properties:
      coins:
        type: integer
      completedSteps:
        type: integer
      id:
        type: integer
      lastLevelTitle:
        type: string
      totalSteps:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  internal_transport_http.upsertDialogInput:
    properties:
      steps:
        items:
          type: integer
        type: array
    type: object
host: localhost:5000
info:
  contact: {}
  description: API for web game for learning Chinese
  title: Chinese Game API
paths:
  /auth/login:
    post:
      description: Create a JWT token for user
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.signUpInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      description: Create a new user account with student role
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.signUpInput'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /level:
    get:
      description: Retrieve a list of all levels, optionally including their steps
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chinese-game-backend_internal_domain.Level'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Get all levels
      tags:
      - levels
  /level/{id}:
    get:
      description: Retrieve a specific level by its ID, including its steps
      parameters:
      - description: Level ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chinese-game-backend_internal_domain.Level'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Get level by ID
      tags:
      - levels
  /level/{id}/step:
    post:
      description: Add a new step to a specific level
      parameters:
      - description: Level ID
        in: path
        name: id
        required: true
        type: integer
      - description: Step data
        in: body
        name: step
        required: true
        schema:
          $ref: '#/definitions/chinese-game-backend_internal_domain.LevelStep'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chinese-game-backend_internal_domain.LevelStep'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Create a new step for a level
      tags:
      - levels
  /level/{id}/step/{step_id}:
    delete:
      description: Remove an existing step of a specific level
      parameters:
      - description: Step ID
        in: path
        name: step_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Delete a step for a level
      tags:
      - levels
    put:
      description: Update an existing step of a specific level
      parameters:
      - description: Step ID
        in: path
        name: step_id
        required: true
        type: integer
      - description: Level ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated step data
        in: body
        name: step
        required: true
        schema:
          $ref: '#/definitions/chinese-game-backend_internal_domain.LevelStep'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Update a step for a level
      tags:
      - levels
  /progress:
    post:
      description: Mark a step as completed for the authenticated user
      parameters:
      - description: Step ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.completeStepInput'
      responses:
        "204":
          description: Step marked as completed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user progress for a specific step
      tags:
      - Progress
  /progress/levels:
    get:
      description: Retrieve a list of level IDs that the authenticated user has completed
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get completed levels for the authenticated user
      tags:
      - Progress
  /progress/levels/{level_id}:
    get:
      description: Determine if the authenticated user has completed a specific level
      parameters:
      - description: Level ID
        in: path
        name: level_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check if a level is completed for the authenticated user
      tags:
      - Progress
  /step/{step_id}/dialog:
    put:
      description: Create or update the dialog for a specific step
      parameters:
      - description: Step ID
        in: path
        name: step_id
        required: true
        type: integer
      - description: Dialog steps
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.upsertDialogInput'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      summary: Upsert dialog for a step
      tags:
      - levels
  /teacher/invite-code:
    get:
      description: Retrieve the invite code for the authenticated teacher
      responses:
        "200":
          description: Invite code
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the invite code for the authenticated teacher
      tags:
      - User
  /teacher/students:
    get:
      description: Retrieve a list of students associated with the authenticated teacher
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_transport_http.studentProgressResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get students associated with the authenticated teacher
      tags:
      - User
  /user/join:
    post:
      description: Join a student to a teacher using an invite code
      parameters:
      - description: Invite Code
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.joinStudentInput'
      responses:
        "204":
          description: Student successfully joined the teacher
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join a student to a teacher using an invite code
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"